Seted in a separate file the tests of the Passcode
